+++ src/lib/components/guide/DemoItem.svelte
