+++ src/lib/components/guide/GuideNav.svelte
