+++ src/lib/components/guide/StepByStep.svelte
