+++ src/lib/components/guide/VideoGuide.svelte
