+++ src/lib/components/index/GamePage.svelte
