+++ src/lib/components/index/LeaguePage.svelte
