+++ src/lib/components/index/game-page/QueueItem.svelte
