+++ src/lib/components/index/game-page/QueueTeamStats.svelte
