+++ src/lib/components/index/game-page/Reset.svelte
