+++ src/lib/components/index/game-page/TeamManagement.svelte
