+++ src/lib/components/index/game-page/TimeSelector.svelte
