+++ src/lib/components/index/game-page/Timer.svelte
