+++ src/lib/components/index/table-page/LeagueTable.svelte
