--- src/lib/components/main-page/PageTransition.svelte
@@ -1,17 +1,0 @@
-i<script lang="ts">
-	import type { Snippet } from 'svelte';
-
-	// Props
-	let { children }: { children: Snippet } = $props();
-</script>
-
-<div class="page-transition">
-	{@render children()}
-</div>
-
-<style>
-	.page-transition {
-		width: 100%;
-		position: relative;
-	}
-</style>
